LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-construction.crossOriginIsolated.https.any-expected.txt


FAIL Test SharedArrayBuffer constructed I420 VideoFrame assert_equals: plane format expected (string) "I420" but got (object) null
FAIL Test Uint8Array(SharedArrayBuffer) constructed I420 VideoFrame assert_equals: plane format expected (string) "I420" but got (object) null
